TileManagerValidation
Improved validation to use universal function so to make sure bounds are
correct use function ValidateList("Whatever list to validate"), means if
we want we can other types of lanes like railways easier (if we do), The
validation does a couple of other sanity checks like making sure the top
is a higher number than lower. The validation also makes sure that all
lanes are two tiles in size ready for increasing size of spawned
objects.